Integer -> Long

+++ src/main/java/com/oficina/saude/repository/Farmaceuticos.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.oficina.saude.model.Farmaceutico;
 
-public interface Farmaceuticos extends JpaRepository<Farmaceutico, Integer>{
+public interface Farmaceuticos extends JpaRepository<Farmaceutico, Long>{
 
 }
