+++ src/main/java/com/oficina/saude/repository/Medicos.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.oficina.saude.model.Medico;
 
-public interface Medicos extends JpaRepository<Medico,Integer>{
+public interface Medicos extends JpaRepository<Medico,Long>{
 
 }
